Have avifImageSetMetadataExif extract orientation
Have avifImageSetMetadataExif() parse the Exif metadata and extract the
Exif orientation to irot and imir properties.

Remove avifImageExtractExifOrientationToIrotImir() from the public API
but keep it as an internal function.

Call avifRWDataSet(&image->exif, exif, exifSize) directly when we just
need to set image->exif.

Based on the reverse of the following commit of Yannis Guyon's:
https://github.com/AOMediaCodec/libavif/pull/1099/commits/9b631db687834aeb8a52884d109f4529cc69a824