Install meson,cargo,nasm as Debian/Ubuntu packages
The current command used to install cargo is:
  curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y

This command executes a shell script that is not in the source
repository using `curl | bash`. This is a potential security
vulnerability. We should not pipe curl output directly to a shell.

The current command used to install meson is:
  pip3 install meson

Using pip install without manifest files (requirements.txt) and checksum
verification (--require-hashes) is a potential security vulnerability.

Bug: b/248582328
Bug: b/253297622