avifmetadatatest: don't include avif/internal.h
avifmetadatatest can only call public functions because it may be linked
with the libavif shared library (when BUILD_SHARED_LIBS is set to ON).
Replace the call to the internal function
avifImageExtractExifOrientationToIrotImir() with a call to the public
function avifImageSetMetadataExif(), and include avif/avif.h instead of
avif/internal.h.

Fix https://github.com/AOMediaCodec/libavif/issues/1162.