Require min and max quanitizers be both specified
Require the min and max quantizers be both specified or both
unspecified. Since the default values of the min and max quantizers are
not well known (24 and 26 respectively) and may be changed, avifenc
users should not set only one of them. So this new requirement is
unlikely to break many existing scripts.