Re-order the boxes within "stbl"
The ISO-BMFF spec has a "recommendation" for how the boxes within
the "stbl" box should be ordered.

Text from Section 6.2.3 of the spec:
It is *recommended* that the boxes within the Sample Table Box be in the
following order: Sample Description, Time to Sample, Sample to Chunk,
Sample Size, Chunk Offset.

Change the writer code to update the order of these boxes to match
what is recommended in the spec.

Old order: stco, stsc, stsz, stss, stts, stsd.
New order: stsd, stts, stsc, stsz, stco, stss.

This also matches the box ordering used by other tools like ffmpeg.

Motivation behind this change:
The MP4 reader in the Android platform fails if this ordering is not
followed. While that is not the right behavior (since the spec does not
mandate the order, it is simply a "recommendation"), there is no harm
in fixing libavif to produce AVIF animations that are consumable on
Android.

Verification:
* Compliance warden has no issues with the new ordering.
* Files are playing back on Chrome like before.