avifenc: Fix an off by one error
Computation of isImageSequence in avifenc has an off by one error.

Since there is an encode before the while loop, >0 check is incorrect
and it must be >=0. A simpler way to check for image sequence is to
simply see if we enter the while loop or not.

Also, fix a typo in CHANGELOG.md.