Rename the AVIF_UNKNOWN_DURATION64/32 macros
The AVIF_UNKNOWN_DURATION64/32 macros are renamed
AVIF_INDEFINITE_DURATION64/32 because their values (all 1s) are
described as indefinite duration in ISO/IEC 14496-12:2022.
AVIF_UNKNOWN_DURATION64/32 look similar to the
AVIF_REPETITION_COUNT_UNKNOWN macro and could be confusing.

In src/write.c, correct the comments for the fields in the EditListBox
(elst).