Add quality and qualityAlpha to avifEncoder
Change the default of minQuantizer and minQuantizerAlpha to
AVIF_QUANTIZER_BEST_QUALITY (0). Change the default of maxQuantizer and
maxQuantizerAlpha to AVIF_QUANTIZER_WORST_QUALITY (63).

The default of quality is based on the average of minQuantizer and
maxQuantizer. Similarly the default of qualityAlpha is based on the
average of minQuantizerAlpha and maxQuantizerAlpha.

Map quality and qualityAlpha in the range of 0..100 linearly to
quantizer values (QP) in the range of 63..0.