Refactor YUV->RGB conversion with libyuv
Rewrite the YUV to RGB conversion functions that use libyuv. Split the
code into two logical pieces: function pointer selection and function
pointer application.

The function pointer selection contains the logic to select the
appropriate libyuv function pointer for all the bitdepths. In the
function pointer step, we simply apply whichever function pointer was
selected in the prior step (downshifting to 8bit if necessary).

This makes the code simpler to parse visually and avoid some redundant
computations.

This is simply a refactoring of code. The functionality should remain
unchanged.