Call color config aom_codec_control on first image
The color config is part of the AV1 sequence header OBU, so the
aom_codec_control() calls related to the color config only need to be
called before encoding the first image.

Also if a value is the default in libaom, then skip the
aom_codec_control() call. aom_codec_control() does a lot of extra work
in addition to setting a config parameter, so we should try to skip
unneeded aom_codec_control() calls.