Add targetSize in avifEncoder
targetSize is an alternative to quality and qualityAlpha for setting
the intended rate-distortion of the encoded image.
It uses a binary search over quality and/or qualityAlpha to find the
settings leading to the encoded image with a size in bytes as close to
the given targetSize as possible.

It is imprecise because a lower quality does not always generate a
smaller encoded payload, and because the input granularity is limited
by the quantizer setting, which has only 64 possible integer values.