Add support for threaded YUV to RGB conversion
Add a new function avifImageYUVToRGBThreaded which takes a
maxThreads parameter and potentially uses up to that many threads
to parallelize the conversion.

If there are n jobs, each job works on a portion of the image in
parallel (each job will work on an image of size
width x (height / 4)).

Microbenchmarking shows that the function scales almost linearly
with the number of threads on larger images (2k or above). For
smaller images, performance is either neutral or slightly better.