Fix image8 in avifImageYUVToRGBLibYUV()
The condition at the end of avifImageYUVToRGBLibYUV() that tests
if (!image->alphaPlane || !image->alphaRowBytes) was potentially using
a pointer to image8 which was no longer on the stack. Fix this issue
detected by sanitizers by using a temporary image_downshifted_to_8bit
variable (replacing the previous image8 one) and a new pointer image8
that either points to image or to image_downshifted_to_8bit. The image
pointer is no longer modified, image8 is used as libyuv function args
now.

Free all planes in image_downshifted_to_8bit and not only alpha to
avoid a leak when downshiftAlpha is true.