avifPNGRead: Don't add alpha filler if no alpha
If the PNG input has no alpha, don't add fully-opaque alpha filler.
Although avifEncoderAddImage() ignores a fully-opaque alpha plane, it
does so only when encoding a single image. When we encode a progressive
image, the current implementation encodes the layers as potentially
different images, so the avifImageIsOpaque() check for a single image is
skipped and the alpha plane is encoded.