avifenc: Add an experimental --progressive option
If --progressive is specified, avifenc will encode a progressive,
layered image. The current implementation is hardcoded to use two
layers.

The --progressive option is marked as experimental because it doesn't
work properly even though it produces a correctly encoded layered image.

Refactor avifEncodeImagesFixedQuality(). Move the big while loop in
avifEncodeImagesFixedQuality() to a new avifEncodeRestOfImageSequence()
function. Then change avifEncodeImagesFixedQuality() to call either
avifEncodeRestOfImageSequence() or avifEncodeRestOfLayeredImage()
depending on whether we are encoding an image sequence or a progressive,
layered image.