Track which images steal planes from codecs
When we call codec->getNextImage() or destroy a codec, the planes stolen
from the codec will become invalid. Track which images steal planes from
a codec, and set their planes to NULL before we call
codec->getNextImage() or after we destroy the codec.

This change would have caused avifgridapitest and avifincrtest to fail
and caught b/278249578.