Don't test potentially dangling pointers
In avifDecoderDataCopyTileToImage(), if a shared codec is used to decode
a grid image, the yuvPlanes pointers of firstTile->image will become
invalid when the shared codec decodes the second tile. Although we only
test if the pointers are null and do not dereference them, it is best to
avoid even testing them. Also, the yuvFormat check should cover this
check.