android_jni: Use specific files for StillImageTest
Instead of iterating over the assets directory, used a fixed set
of files. This will let the tests assert for dimensions and
depth.  This will also prevent the tests from accidentally
passing in case the assets directory is not included correctly.

Also introduce a utility class to refactor some of the common
test code.