workflows: Run android emulator tests on CI
Add a workflow that runs android emulator tests on CI. It is set
up to run only on changes to the android_jni/ subdirectory since
it's a bit slow (at least 15 minutes).

Add a note to android_jni/README.md about the maven hosted
version of the library (this forces the workflow to run on this
change).