Fix gamma for some PNG files.
If the PNG file contains a gAMA chunk and a cHRM chunk, and does not
contain an iCCP chunk or an sRGB chunk, then convert the RGB values
from the gamma specified in the gAMA chunk to the sRGB gamma curve.

Partially fixes #1139 since the cHRM chunk itself is still ignored.