android_jni: Handle JNI exceptions
When calling JNI functions, the default behavior is that the native
code execution will continue and a pending exception will be queued
to be thrown when the control goes back to the Java layer.

This is not desirable when multiple JNI calls are being made and the
second one depends on the successful execution of the first. Suppress
JNI level exceptions and simply return an error from those functions
instead of throwing a JNI exception in the java layer. This also
avoids some SIGABRTs in the native code.

Also, add nullptr checks for return values of FindClass and
GetFieldID.