Force keyframe for alpha if color is a keyframe
When the encoder supports it (as of now only libaom does),
disable lagged output when encoding animated images with alpha
channel. When the utput of the color planes is a keyframe, force
a keyframe on the alpha channel as well.

While the spec does not enforce this, and the libavif decoder can
handle files without this requirement just fine, this change is
still an enhancement because most decoders look at whether or not
the color frame is a keyframe to determine seek points.

Fixes: Issue #841