Write primaries and transfer characteritics info in decoded png.
If there is no ICC but the cicp values are different from sRGB,
they are added to the png file using the cHRM, gAMA and cICP chunks.

The cHRM and gAMA chunks are preferred since they are older and
more widely understood (although some viewers may ignore them).
The new cICP chunk is added if gAMA is not enough to specify the
transfer characteristics.