Experimental reduced HEIF header
Add AVIF_ENABLE_EXPERIMENTAL_AVIR flag.
Remove a few hundred bytes of container boilerplate for simple images.
This prototype uses a small box that expands in meaning to a full meta
box with all defined items and properties.

Warning: This feature is experimental and forbidden by the current
AVIF specification.