Add support for reading PNG gAMA, cHRM and sRGB chunks (#1448)
This implementation tries to match gAMA and cHRM values with CICP values and set TC and CP if a successful match is found, otherwise generates an ICC profile based on gAMA and cHRM values.

This implementation also handles sRGB chunk by setting sRGB CP and TC.

The gamma=2.2 and BT709 primaries case, which is usually used to tag images as sRGB, is not specially treated as sRGB, to align with how browsers handle such PNG images.

Introduc --ignore-profile flag to ignore any color profile info including ICC profile and PNG gAMA, cHRM, sRGB chunks. --ignore-icc is considered deprecated and treated as an alias of --ignore-profile.