avifenc.c: Remove the --longhelp option
Go back to using the -h or --help option to request the full syntax
help. The short syntax help is only used as the error message for
certain kinds of command-line errors.

This follows the design of libwebp2/examples/cwp2.cc.

The rationale is that when users type the -h or --help option, they are
most likely interested in the full syntax help, so it doesn't make sense
to require an intermediate step to get there.

We can revisit this design if we add a "basic" syntax help message that
is more informative than the current short syntax help message.

Bug: b/275100954