reformat_libyuv: Add ignore CFI annotation
libyuv is a C++ library and defines custom types (struct, enum, etc) in
a namespace when conditionally compiled with a C++ compiler. When
accessed from a C library like libavif, via a function pointer, this
leads to signature mismatches in the CFI sanitizers since the C++ header
file defines the types within the namespace and the C code has the types
without the namespace. In order to avoid this CFI error, we tag some
functions as an exception when being compiled with CFI enabled.

For more details on clang's CFI see:
https://clang.llvm.org/docs/ControlFlowIntegrity.html.

For a simpler example of this bug, please see:
https://github.com/vigneshvg/cpp_c_potential_cfi_bug

Sample CFI failure in Chromium: crbug.com/1473228