Change imir.mode back to imir.axis
Change imir.mode back to imir.axis for the libavif 1.0.0 release.

In ISO/IEC 23008-12:2017 Draft Amendment 2, the 'axis' field of the
'imir' property was renamed 'mode', so we changed imir.axis to
imir.mode. But in ISO/IEC 23008-12:2022, the 'axis' field of the 'imir'
property was not renamed.

In comments and messages, avoid using the words "vertical" and
"horizontal" to describe how the mirroring is performed. Also avoid the
word "axis" when possible.

Code that needs to support both old and new versions of libavif will
need to test #if AVIF_VERSION_MAJOR >= 1.