Add golden tests for the written AVIF metadata structure.
Depends on MP4Box (from the gpac project) to dump an AVIF file's
metadata as XML.

This test catches changes in write.c that may or may not be intentional.
On failure, it prints instructions on how to update the goldens if needed.
The test is disabled by default, but enabled in a new CI workflow.