Print gain map info in avifenc/avifdec. (#1598)
In avifenc, the gain map's detailed info like pixel size, yuv format etc. are printed.
In avifdec, only the presence/absence of a gain map in the AVIF file is printed. 
The full info is not available because enableDecodingGainMap is disabled by default,
and it doesn't make sense to enable it since the gain map won't be savec anyway
(saving to jpeg+gainmap is not implemented).