Allow creating images with a different grid for the gain map.
The main image can have a different number of cols/rows than
the gain map, or one can be a grid and the other be a single image.
Also allow setting differnet values for gain map metadata version
fields.
Add code in gainmaptest.cc to generate test images. These images
are used in tests and as seeds by the fuzzer.