fix: unused parameter error in third_party libyuv with release build
In several functions in libyuv/scale.c, the only reference to the
parameter `src_width` is an assert statement. When NDEBUG is defined (as
it is on some platforms when compiling with CMAKE_BUILD_TYPE=Release)
the assert condition is removed from the source and an unused-parameter
warning/error is raised.