Patch alpha_video.avif to fix the iref box
The iref box contains an invalid item_id of 0. This file started
failing since `1fbea62`.

This file has always had this issue but it was ignored because it is
an animated AVIF and the tracks information is correct. If we try to
decode it as a still image by setting `requestedSource` to primary
item, then avifdec (prior to `1fbea62`) decodes it as an image with
alpha (which is not correct).

Patching the file fixes the Android CI failures on this file.