android_jni: Support scaling (#1675)
Add support for scaling in the android_jni wrapper now that scaling is supported natively in libavif.
This does not add any new API functions. Merely pass a bitmap of a different size to scale the output
as desired. It works for both the utility class usage as well as the regular class usage.

---------

Co-authored-by: Vignesh Venkatasubramanian <vigneshv@google.com>