Set some dependency options to be multi-choice.
The current options are complex:
- you want a local AOM, you need to set to variables (AVIF_CODEC_AOM
and AVIF_LOCAL_AOM to ON)
- they can be confusing: some work in best effort, e.g. if you do
not AVIF_LOCAL_LIBYUV, it tries to find libyuv if not, oh well, too
bad let's move on.
- you cannot disable some options: if libyuv is found, use it! Oh
wait, I actually never asked for it (causing issues like
https://github.com/AOMediaCodec/libavif/issues/1732)

That's why I propose to set dependency options to:
OFF: it will never be used whether it is findable or in ext
LOCAL: it has to use the version in ext/ and will fail if not found
SYSTEM: it will find_package it and fail if not found

An AUTO version (that uses a dependency if found) should not be
added as it will create more confusion for developers or
package building.