Do not store potentially invalid pointers
Calling avifMetaFindItem() could invalidate all the
existing item pointers that are being stored by the caller (since
the function could resize the item array).

This patch fixes avifDecoderReset by updating the locally stored
pointer after call to avifMetaFindItem.

This is a manual cherry-pick of commit `2041109` on to the v1.0.x
branch.