Allocate the threadData array directly
The threadData array (formerly named tdArray) has a fixed size (i.e.,
the array does not grow or shrink while in use), so we can allocate the
array directly. It is not necessary to define the array as avifArray.

(cherry picked from commit c74189f9123bc01139ed316970f673394a57e814)