Further cleanup in avifParseSampleDescriptionBox()
Rename the `remainingBytes` variable as `sampleEntryBytes`. The
`remainingBytes` name made sense when the variable stored the return
value of avifROStreamRemainingBytes(&s). The variable now stores the
value of sampleEntryHeader.size, so the name `sampleEntryBytes` is more
appropriate.

Remove an extraneous pair of parentheses around a conditional
expression.

Replace an instance of sampleEntryHeader.size with `sampleEntryBytes`.