Fix size 0 handling in avifParse()
header.size in avifParse is not the raw value that is found in
the input stream. It is calculated in avifROStreamReadBoxHeaderPartial.

So when handling size == 0 in avifParse(), we cannot use header.size
since it may be 0 in legit cases (where the box itself is empty).

This PR adds a new member to avifBoxHeader which stores the actual
raw size found in the input stream and uses that in avifParse().

Also handle the isSizeZeroBox in avifPeekCompatibleFileType.