Revert "Move gainmap function definitions to gainmap.c"
This reverts commit 4a4ff640088063fdf0d33164f6aa380777492f5a.

The code that was originally in src/gainmap.c is not used by Chrome.
With this commit, Chrome will need to include src/gainmap.c, which in
turn requires src/colrconvert.c.

Let's try again to figure out the reason for the strange unresolved
symbol linker error in a shared library build on Windows, and fix it
properly.