Provide both avifincrtest_helpers{_internal}
Provide both the avifincrtest_helpers and avifincrtest_helpers_internal
OBJECT libraries. This allows us to link avifincrtest with only the avif
library but not with the avif_internal static library. avifincrtest uses
only public API functions of the libavif library, so it does not need to
depend on the avif_internal static library.

Fix https://github.com/AOMediaCodec/libavif/issues/2325.