Fix Clang's -Wformat warnings
Use the %u format specifier with uint32_t arguments.

Cast an enum type to int and use the %d format specifier with it. The
underlying type of an enum type is implementation-defined. GCC uses
unsigned int whenever possible, whereas MSVC always uses int. So there
is no portable format specifier for an enum type. Since all enum
constants have the int type (this is not implementation-defined), it is
safe to cast an enum type to int and use the %d format specifier with
it.