libjpeg: Pass -DWITH_CRT_DLL=ON to cmake
On Windows, add
  target_link_options(avifenc PRIVATE /WX)
to CMakeLists.txt to treat linker warnings as errors. This pull request
fixes the following linker warning:
  warning LNK4098: defaultlib 'LIBCMT' conflicts with use of other libs;
  use /NODEFAULTLIB:library

Adding /NODEFAULTLIB:LIBCMT to the target_link_options() call, as the
warning message suggests, is another way to fix the linker warning.

For more info on the WITH_CRT_DLL cmake option, see
https://github.com/libjpeg-turbo/libjpeg-turbo/issues/573.