Fix overflows in avifRGBImageAllocatePixels()
Calculate rowBytes in uint32_t. Only the allocation size needs to be
size_t.

Also make sure it is safe to cast various rowBytes fields to ptrdiff_t.
We need to do this when subtracting rowBytes from a pointer to go back
one row.

Part of the fix to https://github.com/AOMediaCodec/libavif/issues/2271.