Check if it is safe to call libyuv or libsharpyuv
The width, height, and stride parameters of libyuv and libsharpyuv
functions are all of the int type. The corresponding fields in libavif
are of the uint32_t type. So we need to check if it is safe to pass the
libavif uint32_t values to libyuv and libsharpyuv.

Related to https://github.com/AOMediaCodec/libavif/issues/2271.