Remote check that the meta box is zero when searching for a tmap.
Whether the meta box has size zero or not, it's read in one go.
So if the meta box has been seen, all of it has been seen.
The only case where it could be read partially is if the provided
sizeHint value is too small, but then the parsing code still assumes
it has read the whole box.
If the whole box is not available yet, the reading function is
supposed to return AVIF_RESULT_WAITING_ON_IO.