Disable the MSVC C4232 warning
The way we use the addresses of the libyuv functions in
src/reformat_libyuv.c, the addresses of those functions do not need to
be "static", which I think means fixed from execution to execution. As
long as the addresses of those functions are fixed during an execution,
that is fine for src/reformat_libyuv.c.

Fix https://github.com/AOMediaCodec/libavif/issues/2447.