Update parseAV2SequenceHeader() for AVM CWG-E103
In https://gitlab.com/AOMediaCodec/avm/-/merge_requests/1183, the syntax
elements for the max frame width and height and color_config() were
moved to the beginning of the sequence header OBU.
parseAV2SequenceHeader() needs to be updated if AVM has this change.
Since this change has not appeared in an AVM release tag, add the
AVIF_ENABLE_CWG_E103 macro to indicate whether AVM has this change.

To facilitate the changes to parseAV2SequenceHeader(), split
parseSequenceHeaderProfile() into two functions:
  - parseSequenceHeaderProfile() parses only seq_profile
  - parseSequenceHeaderLevelIdxAndTier parses seq_level_idx and seq_tier
and change parseSequenceHeaderColorConfig() to not parse
separate_uv_delta_q.