Handle animations with audio tracks.
HEIF allows animations to have audio tracks, subtitle tracks etc.
Previously, these files would yield and error because the track header
for these tracks has width/height of zero. Now the size is only checked
for tracks with a known video handler type ('pict', 'vide', 'auxv').

Moreover, we now only use tracks with a handler type of 'pict'
for the main track (as specified in HEIF). Previously, the handler
type was not checked.