Revert to previous default behavior.
By default, only the first frame is decoded. --index all must be used
explicitly to decode all frames. In that case, the first frame is no longer
written twice, only the -xxx files are written.
Use ten padding zeros for the frame number, to accomodate the max value of an int32.
A fixed number of leading zeros is easier for scripts to deal with.