codec_avm.c: Removed unneeded code for options
All versions of AVM that we need to support have the
aom_codec_set_option() function. Remove the code that assumes
aom_codec_set_option() is not present.

This also fixes a bug in avifProcessAOMOptionsPostInit() that
incorrectly concatenates the two conditionally-compiled pieces of code
when the file was copied from codec_aom.c and the
HAVE_AOM_CODEC_SET_OPTION macro was removed.