avifrgbtoyuvtest: Do not compute diff_sum
The sum of pixel sample differences is not a good measure of distortion
because it does not satisfy all intuitive properties of the length of a
vector (the mathematical term is the norm of a vector). Do not compute
diff_sum.

Rename max_abs_average_diff to max_average_abs_diff and use it as the
maximum of the average absolute value of pixel sample difference.