avifenc.c Use better formula in quantizerToQuality
The original formula was apparently derived by inverting the
quality-to-quantizer conversion formula in src/write.c. That derivation
is not ideal because it included the `+ 50` rounding term. It is better
to omit the `+ 50` rounding term, invert the formula, and then add a
`+ 31` rounding term.

The new formula is already used in avifenc.c.