Revert "Set Unspecified CICP in Seq Header OBU (#2687)"
This partially reverts commit fa089a0c3f8ed6e202034589cf98394e0226d66c.
The added test avifcicptest.cc is not reverted.
AVM is also left as is, as it's a new codec and there are no backward
compatibility concerns.
Set the CICP in codec_svt.c which was not done previously.

Fixes #2850