Add support for Apple gain maps (iPhone jpegs)
These use a different XMP syntax on the gain map image, and do not
carry XMP on the main image.
Moreover, the alternate headroom may not be present in the XMP, in which case it's read
from the Apple Maker Notes in Exif.

Add a --yuv-gain-map to set the gain map pixel format (chrome subsampling).
This is useful because Apple gain maps are grayscale, and Chrome currently
has a bug where grayscale gainmaps are ignored https://g-issues.chromium.org/issues/472455759